refactor(test): split session stop integration coverage
